[Decomp][Player] Cohesion split: extract stat + skill formulas to PlayerStats.cpp
Moves the contiguous combat-rating, base-mod, and skill-formula block
out of Player.cpp into a new file under src/game/Object/.

Player.cpp: 17,702 -> 16,496 lines (-1,206).
PlayerStats.cpp: 1,289 lines, 35 methods.

Methods moved (combat ratings + base mods):
- HandleBaseModValue, GetBaseModValue, GetTotalBaseModValue
- GetShieldBlockDamageValue
- GetMeleeCritFromAgility, GetDodgeFromAgility, GetParryFromStrength,
  GetSpellCritFromIntellect
- GetRatingMultiplier, GetRatingBonusValue, GetExpertiseDodgeOrParryReduction
- OCTRegenMPPerSpirit
- ApplyRatingMod, UpdateRating, UpdateAllRatings, SetRegularAttackTime

Methods moved (skill formulas + accessors):
- UpdateSkill, UpdateCraftSkill, UpdateGatherSkill, UpdateFishingSkill,
  UpdateSkillPro
- ModifySkillBonus, UpdateSkillsForLevel, UpdateSkillsToMaxSkillsForLevel
- SetSkill, HasSkill, GetSkillStep
- GetSkillValue, GetMaxSkillValue, GetPureMaxSkillValue,
  GetBaseSkillValue, GetPureSkillValue
- GetSkillPermBonusValue, GetSkillTempBonusValue

No header change; no signature change; no behaviour change.
